+++ Meta/check-style.py
@@ -24,7 +24,7 @@ LICENSE_HEADER_CHECK_EXCLUDES = {
     'AK/Checked.h',
     'AK/Function.h',
     'Userland/Libraries/LibJS/SafeFunction.h',
-    'Userland/Libraries/LibC/elf.h',
+    'Userland/Libraries/LibELF/ELFABI.h',
     'Userland/Libraries/LibCodeComprehension/Cpp/Tests/',
     'Userland/Libraries/LibCpp/Tests/parser/',
     'Userland/Libraries/LibCpp/Tests/preprocessor/'
