+++ Userland/Libraries/LibELF/DynamicObject.h
@@ -13,8 +13,8 @@
 #include <AK/DeprecatedString.h>
 #include <AK/RefCounted.h>
 #include <Kernel/Memory/VirtualAddress.h>
-#include <LibC/elf.h>
-#include <LibC/link.h>
+#include <LibELF/ELFABI.h>
+#include <link.h>
 
 namespace ELF {
